Revert "fix the redirect of "View All" button in threat dashboard"
