+++ apps/dashboard/web/polaris_web/web/src/apps/main/LocalStorageStore.js
@@ -2,7 +2,7 @@ import {create} from "zustand"
 import { devtools, persist } from "zustand/middleware"
 
 import pako from "pako"; // Gzip Compression
-import { createGzipStorage } from "./gzipStorage";
+import { createGzipStorage } from "./PersistStore";
 
 // Custom Storage with Gzip Compression for localStorage
 const gzipLocalStorage = createGzipStorage(localStorage);
@@ -14,9 +14,6 @@ const initialState = {
     defaultIgnoreSummaryTime: 2 * 60 * 60,
     lastEndpointEpoch: 0,
     agenticNewLayout: false,
-    // Mirrors PersistStore.dashboardCategory's old default logic, kept inline (not imported
-    // from labelHelper.js) to avoid a circular import with PersistStore.js at module init time.
-    dashboardCategory: window.DASHBOARD_CATEGORY || "API Security",
 };
 
 let localStore = (set) => ({
@@ -67,14 +64,6 @@ let localStore = (set) => ({
             set({ agenticNewLayout });
         } catch (error) {
             console.error("Error setting agenticNewLayout:", error);
-        }
-    },
-
-    setDashboardCategory: (dashboardCategory) => {
-        try {
-            set({ dashboardCategory });
-        } catch (error) {
-            console.error("Error setting dashboardCategory:", error);
         }
     },
